panicmode: Avoid pushing the nonterminal string instead of its attribute
Instead of pushing the string representation of the valid nonterminal,
push its attribute. It won't be possible to push its attribute since it
is not know at the time. One workaround is to push a dummy node instead
with the same type as that of the actual attribute (which will be used
when reduced). Since we are only concerned with error reporting and not
the evaluations done after reductions, this approach will work.

The problem which now remains is to guess the type of the attribute. The
following approaches come to me on first look:
  - Use an empty interface??
    Seems like no -- https://stackoverflow.com/q/27971895/5107319
  - Use the data in "productionstable.go" to get an idea of the type
    (quite hacky).

For the purposes of a demo using the BNF file of "calc" example
available in gocc, the type of dummy node is int64 just to make things
work.

Update the Makefile and test of calc example to demonstrate panic mode.

Demo
~~~~
On executing `go test`, it should be noted that parsing doesn't fail
on invalid input. The debug output should give an idea of how the stack
is modified during the run.